+++ openapidocs/mk/generate.py
@@ -9,9 +9,9 @@ def generate_document(source: str, destination: str, style: Union[int, str]):
     # parameter in this function
 
     data = read_from_source(source)
-    handler = OpenAPIV3DocumentationHandler(data, style=style)
+    handler = OpenAPIV3DocumentationHandler(data, style=style, source=source)
 
-    html = handler.write(data)
+    html = handler.write()
 
     # TODO: support more kinds of destinations
     with open(destination, encoding="utf8", mode="wt") as output_file: