+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2022-05-08
+- Adds support for OpenAPI specification files split into multiple files
+  https://github.com/Neoteroi/mkdocs-plugins/issues/5
+- Adds support for `externalDocs` and `tags` root properties
+
 ## [1.0.1] - 2022-05-05
 - Adds a new output style, to provide an overview of the API endpoints with
   PlantUML